projects/streaming-kafka-spark/requirements.txt
kafka-python==2.0.2
pyspark==3.3.2
psycopg2-binary==2.9.9
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
jsonschema==4.19.2
requests==2.31.0
py4j==0.10.9.5
